- fixed speed calculations not working anymore due to terran dlc - added income per hour column to wares window

+++ X4_Editor/UIModel/UIModelProjectile.cs
@@ -126,8 +126,8 @@ namespace X4_Editor
             }
         }
 
-        private int m_Speed;
-        public int Speed
+        private double m_Speed;
+        public double Speed
         {
             get { return m_Speed; }
             set
